Aktualisiert

+++ Diga.Core.Api.Win32/NmHdr.cs
@@ -8,6 +8,6 @@ namespace Diga.Core.Api.Win32
     {
         public IntPtr hwndFrom;
         public uint idFrom;
-        public uint code;
+        public long code;
     }
 }
